feat: encode pruning tables in binary for smaller file size

+++ include/pruning.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <cstdint>
+
 void writePruningTable(int *pruningTable, int tableSize, const char *fileName);
-void loadPruningTable(int *table, int tableSize, const char *fileName);
+void loadPruningTable(uint8_t *table, int tableSize, const char *fileName);
